+++ CHANGELOG.md
@@ -6,14 +6,16 @@ The following changes have been implemented but not released yet:
 
 ## [Unreleased]
 
+The following changes have been released:
+
+## [0.3.1] - 2021-10-21
+
 ### Bugfix
 
 - The property for access issuer discovery has been aligned with the one used on
   the server side, i.e. `http://www.w3.org/ns/solid/terms#accessIssuer`.
 - The discovery for the access issuer no longer relies on the Pod server, but rather
   on the UMA server.
-
-The following changes have been released:
 
 ## [0.3.0] - 2021-10-15